+++ src/components/organisms/OrderHistorySection/OrderHistorySection.tsx
@@ -1,6 +1,6 @@
 ﻿import { ShoppingBag } from "lucide-react";
 import { useCallback, useEffect, useState } from "react";
-import { getOrderHistory } from "../../../api/order";
+import { getOrderHistory } from "../../../api/orders";
 import type {
   OrderHistoryItem,
   OrderHistoryPeriod,
